pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hospital-portal-system"
version = "0.1.0"
description = "A Hospital Management System built with Flask"
requires-python = ">=3.11,<3.12"
dependencies = [
    "flask==3.1.1",
    "flask-sqlalchemy==3.1.1",
    "flask-login==0.6.3",
    "flask-migrate==4.0.5",
    "flask-wtf==1.2.1",
    "python-dotenv==1.1.0",
    "email-validator==2.2.0",
    "alembic==1.13.1",
    "sqlalchemy==2.0.30",
    "werkzeug==3.1.3",
    "jinja2==3.1.3",
    "wtforms==3.1.2",
    "blinker==1.9.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "greenlet==3.0.3",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "mako==1.3.2",
    "markupsafe==2.1.5",
    "typing-extensions==4.14.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy"
]

[tool.hatch.build.targets.wheel]
packages = ["hospital_flask"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 